refactor(lockfile): split drift checks
